# projects/neural-sync/artifacts/api-contracts.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEURAL SYNC API",
    "version": "1.0.0",
    "description": "Multi-dimensional AI-driven developer–project compatibility engine. All protected routes require JWT Bearer authentication per ADR-002. The 500 ms p95 latency SLA applies to the synchronous POST /matches response per ADR-003. LLM explanations are generated asynchronously (Phase-1: Google Gemini; Claude API is a supported alternate per ADR-004); the initial response always contains a deterministic stub (≥50 chars) per ADR-003/ADR-004. Error envelope on all 4xx/5xx: {error_code, message, request_id}.",
    "contact": {
      "name": "NEURAL SYNC Engineering",
      "email": "engineering@neural-sync.example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server"
    },
    {
      "url": "https://neural-sync.example.com/api/v1",
      "description": "Production server"
    }
  ],
  "security": [
    { "BearerAuth": [] }
  ],
...
  "paths": {
...
    "/developers": {
      "post": {
        "tags": ["developers"],
        "summary": "Ingest a new developer profile",
        "description": "Creates a DeveloperProfile. All required fields are validated (skills, work_style vector len=8, motivation_vector len=8, IANA timezone, etc.). Server generates a UUID if id is omitted. Enqueues embedding generation as a background task; the profile is immediately accessible with embedding_status='pending'.",
        "operationId": "createDeveloper",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeveloperProfileCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Developer profile created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperProfile" }
              }
            }
          },
          "400": {
            "description": "Invalid profile data (e.g. work_style vector length ≠ 8)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized — valid JWT required",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "409": {
            "description": "Developer with provided id already exists",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Pydantic validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/developers/{developer_id}": {
      "get": {
        "tags": ["developers"],
        "summary": "Get a developer profile by ID",
        "description": "Returns the full DeveloperProfile. Developers (role=developer) may only retrieve their own profile; managers may retrieve any profile. Returns HTTP 404 if the developer was never created or was erased via DELETE.",
        "operationId": "getDeveloper",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Developer profile found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperProfile" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — developer role accessing another developer's profile",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found (never created, or erased via GDPR DELETE)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error_code": "NOT_FOUND",
                  "message": "Developer with id '3fa85f64-5717-4562-b3fc-2c963f66afa6' not found",
                  "request_id": "req_abc123"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["developers"],
        "summary": "Replace a developer profile",
        "description": "Full replacement of a DeveloperProfile. All required fields must be present. Re-enqueues embedding generation on success. Developers may only update their own profile.",
        "operationId": "updateDeveloper",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DeveloperProfileCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Developer profile updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperProfile" }
              }
            }
          },
          "400": {
            "description": "Invalid profile data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "422": {
            "description": "Validation error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      },
      "delete": {
        "tags": ["developers"],
        "summary": "GDPR right-to-erasure: full cascade deletion",
        "description": "Performs atomic cascade deletion within a single PostgreSQL transaction: (1) DeveloperProfile row, (2) developer_embeddings rows, (3) all MatchRecord rows, (4) all FeedbackRecord rows, (5) all AllocationRecord rows, (6) ExplanationCache entries keyed to the developer. Creates an ErasureAuditLog entry. On transaction failure, rolls back completely and returns HTTP 500 — developer data remains intact. A subsequent GET /developers/{developer_id} returns HTTP 404 after successful erasure.",
        "operationId": "deleteDeveloper",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" }
        ],
        "responses": {
          "204": {
            "description": "Developer fully erased — GDPR compliant. No response body."
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden — only the developer themselves or manager role may request erasure",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "500": {
            "description": "Erasure transaction rolled back — developer data is intact. Check erasure audit log.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error_code": "ERASURE_FAILED",
                  "message": "Cascade deletion rolled back. Developer data is intact. erasure_request_id: 'abc-123'",
                  "request_id": "req_xyz789"
                }
              }
            }
          }
        }
      }
    },
    "/developers/{developer_id}/risk": {
      "get": {
        "tags": ["developers"],
        "summary": "Get burnout and bench risk scores for a developer",
        "description": "Returns burnout_risk_score (>0.6 when ≥48 consecutive weeks at workload_intensity ≥0.8) and bench_risk_score (>0.7 when current project ends within 28 days with no follow-on allocation). Scores computed from AllocationRecord history. Raw behavioral vectors are never exposed.",
        "operationId": "getDeveloperRisk",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" }
        ],
        "responses": {
          "200": {
            "description": "Risk scores returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RiskResponse" },
                "example": {
                  "developer_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                  "burnout_risk_score": 0.72,
                  "bench_risk_score": 0.15,
                  "burnout_risk_badge": "high",
                  "bench_risk_badge": "low",
                  "computed_at": "2026-06-26T10:00:00Z"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
    "/developers/{developer_id}/matches": {
      "get": {
        "tags": ["developers"],
        "summary": "Get ranked project recommendations for a developer",
        "description": "Returns top-K MatchRecord objects ranked by match_score descending. Used by the Developer Dashboard to render recommended project cards with explanation and risk tags. Developers may only query their own matches.",
        "operationId": "getDeveloperMatches",
        "parameters": [
          { "$ref": "#/components/parameters/DeveloperIdPath" },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 },
            "description": "Maximum number of recommendations to return"
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
            "description": "Filter results to match_score ≥ min_score"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked match recommendations",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeveloperMatchesResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "403": {
            "description": "Forbidden",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          },
          "404": {
            "description": "Developer not found",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from POST /auth/login. Algorithm: HS256. Payload: {sub: user_id, role: 'developer'|'manager', iat, exp}. TTL: 1h. Secret loaded from env NEURAL_SYNC_JWT_SECRET. Per ADR-002: store in memory (not localStorage) on the React frontend; refresh token stored in HttpOnly cookie."
      }
    },
    "parameters": {
      "DeveloperIdPath": {
        "name": "developer_id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" },
        "description": "Developer unique identifier (UUID v4)"
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error envelope returned on all 4xx and 5xx responses.",
        "required": ["error_code", "message", "request_id"],
        "properties": {
          "error_code": {
            "type": "string",
            "description": "Machine-readable error code. Examples: NOT_FOUND, INVALID_WEIGHTS, VALIDATION_ERROR, ERASURE_FAILED, SERVICE_UNAVAILABLE, FORBIDDEN, UNAUTHORIZED, CONFLICT"
          },
          "message": {
            "type": "string",
            "description": "Human-readable description of the error, safe to display in developer tooling"
          },
          "request_id": {
            "type": "string",
            "description": "Unique request trace identifier for correlating logs"
          }
        }
      },
...
      "DeveloperProfileCreate": {
        "type": "object",
        "description": "Request body for creating or replacing a DeveloperProfile. Server generates a UUID if id is omitted.",
        "required": [
          "skills",
          "experience_years",
          "preferred_stack",
          "work_style",
          "motivation_vector",
          "timezone",
          "availability_hours",
          "career_goals",
          "project_history"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional client-provided UUID. Server generates one if omitted."
          },
          "skills": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "description": "Technical skill list, e.g. ['Python', 'FastAPI', 'PostgreSQL', 'Docker']"
          },
          "experience_years": {
            "type": "integer",
            "minimum": 0,
            "description": "Total years of professional software development experience"
          },
          "preferred_stack": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "description": "Preferred technology stack"
          },
          "work_style": {
            "type": "array",
            "items": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
            "minItems": 8,
            "maxItems": 8,
            "description": "8-dimensional work-style behavioral vector; each element ∈ [0.0, 1.0]. Dimensions in order: [collaboration, autonomy, structure, innovation, pace, communication, risk_tolerance, remote_preference]"
          },
          "motivation_vector": {
            "type": "array",
            "items": { "type": "number", "minimum": 0.0, "maximum": 1.0 },
            "minItems": 8,
            "maxItems": 8,
            "description": "8-dimensional motivation vector; each element ∈ [0.0, 1.0]. Dimensions in order: [impact, growth, compensation, stability, creativity, recognition, autonomy, mission_alignment]"
          },
          "timezone": {
            "type": "string",
            "description": "IANA timezone identifier, e.g. 'America/New_York', 'Europe/Warsaw', 'Asia/Singapore'"
          },
          "availability_hours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "description": "Weekly hours available for project work"
          },
          "career_goals": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "description": "Career goal keywords or free-text statements, e.g. ['technical leadership', 'distributed systems', 'AI/ML']"
          },
          "project_history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ProjectHistoryEntry" },
            "description": "Past project engagements (may be empty for new developers)"
          },
          "is_self_reported": {
            "type": "boolean",
            "default": true,
            "description": "True if work_style and motivation_vector are self-reported; false if inferred from external data. Influences UI badge per requirements."
          }
        }
      },
      "DeveloperProfile": {
        "description": "Complete developer profile including server-generated fields.",
        "allOf": [
          { "$ref": "#/components/schemas/DeveloperProfileCreate" },
          {
            "type": "object",
            "required": ["id", "embedding_status", "created_at", "updated_at"],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "Unique identifier (server-generated or client-provided)"
              },
              "embedding_status": {
                "type": "string",
                "enum": ["pending", "ready", "failed"],
                "description": "Status of vector embedding generation. 'pending'=background task queued; 'ready'=embeddings available for ANN search; 'failed'=all 3 retry attempts exhausted (matching degrades to keyword-only)"
              },
              "created_at": { "type": "string", "format": "date-time" },
              "updated_at": { "type": "string", "format": "date-time" }
            }
          }
        ]
      },
...
      "RiskResponse": {
        "type": "object",
        "description": "Developer burnout and bench risk scores. Raw behavioral vectors are never included.",
        "required": ["developer_id", "burnout_risk_score", "bench_risk_score", "burnout_risk_badge", "bench_risk_badge", "computed_at"],
        "properties": {
          "developer_id": { "type": "string", "format": "uuid" },
          "burnout_risk_score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Burnout risk 0.0–1.0. Formula: min(1.0, (consecutive_high_intensity_weeks / 48) × workload_intensity_mean × (1 − motivation_alignment_factor)). Score > 0.6 triggers 'high' badge."
          },
          "bench_risk_score": {
            "type": "number",
            "minimum": 0.0,
            "maximum": 1.0,
            "description": "Bench risk 0.0–1.0. Formula: min(1.0, max(0.0, (28 − days_until_project_end) / 28)). Score > 0.7 triggers 'high' badge."
          },
          "burnout_risk_badge": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "description": "low: score < 0.4; medium: 0.4 ≤ score ≤ 0.6; high: score > 0.6"
          },
          "bench_risk_badge": {
            "type": "string",
            "enum": ["low", "medium", "high"],
            "description": "low: score < 0.4; medium: 0.4 ≤ score ≤ 0.7; high: score > 0.7"
          },
          "computed_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of score computation"
          }
        }
      },
...
    }
  }
}